+++ src/controllers/UsersController.php
@@ -375,15 +375,10 @@ class UsersController extends Controller
 
             $message = UserHelper::getLoginFailureMessage($authError, $e->user);
 
-            return $this->asJson([
-                'success' => false,
-                'message' => $message,
-            ]);
+            return $this->asFailure($message);
         }
 
-        return $this->asJson([
-            'success' => $success,
-        ]);
+        return $this->asSuccess();
     }
 
     /**
@@ -394,16 +389,16 @@ class UsersController extends Controller
         // Passing false here for reasons.
         Craft::$app->getUser()->logout(false);
 
-        if ($this->request->getAcceptsJson()) {
-            $return = [
-                'success' => true,
-            ];
+        $data = [];
 
+        if ($this->request->getAcceptsJson()) {
             if (Craft::$app->getConfig()->getGeneral()->enableCsrfProtection) {
-                $return['csrfTokenValue'] = $this->request->getCsrfToken();
+                $data['csrfTokenValue'] = $this->request->getCsrfToken();
             }
 
-            return $this->asJson($return);
+            return $this->asSuccess(
+                data: $data,
+            );
         }
 
         // Redirect to the login page if this is a CP request
@@ -411,7 +406,10 @@ class UsersController extends Controller
             return $this->redirect(Request::CP_PATH_LOGIN);
         }
 
-        return $this->redirect(Craft::$app->getConfig()->getGeneral()->getPostLogoutRedirect());
+        return $this->asSuccess(
+            data: $data,
+            redirect: Craft::$app->getConfig()->getGeneral()->getPostLogoutRedirect()
+        );
     }
 
     /**
@@ -471,7 +469,7 @@ class UsersController extends Controller
 
         if (empty($errors)) {
             if ($this->request->getAcceptsJson()) {
-                return $this->asJson(['success' => true]);
+                return $this->asSuccess();
             }
 
             $this->setSuccessFlash(Craft::t('app', 'Password reset email sent.'));
@@ -554,20 +552,16 @@ class UsersController extends Controller
         $user->setScenario(User::SCENARIO_PASSWORD);
 
         if (!Craft::$app->getElements()->saveElement($user)) {
-            $errors = $user->getErrors('newPassword');
-
-            if ($this->request->getAcceptsJson()) {
-                return $this->asErrorJson(implode(', ', $errors));
-            }
-
-            $this->setFailFlash(Craft::t('app', 'Couldn’t update password.'));
-
-            return $this->_renderSetPasswordTemplate([
-                'errors' => $errors,
-                'code' => $code,
-                'id' => $uid,
-                'newUser' => !$user->password,
-            ]);
+            return $this->asModelFailure(
+                    $user,
+                    Craft::t('app', 'Couldn’t update password.'),
+                    errorAttribute: 'newPassword'
+                ) ?? $this->_renderSetPasswordTemplate([
+                    'errors' => $errors,
+                    'code' => $code,
+                    'id' => $uid,
+                    'newUser' => !$user->password,
+                ]);
         }
 
         // If they're pending, try to activate them, and maybe treat this as an activation request
@@ -584,13 +578,12 @@ class UsersController extends Controller
 
         if ($this->request->getAcceptsJson()) {
             $return = [
-                'success' => true,
                 'status' => $user->getStatus(),
             ];
             if ($loggedIn && Craft::$app->getConfig()->getGeneral()->enableCsrfProtection) {
                 $return['csrfTokenValue'] = $this->request->getCsrfToken();
             }
-            return $this->asJson($return);
+            return $this->asSuccess(data: $return);
         }
 
         // Can they access the CP?
@@ -1297,20 +1290,11 @@ JS,
                 $user->clearErrors('unverifiedEmail');
             }
 
-            if ($this->request->getAcceptsJson()) {
-                return $this->asJson([
-                    'errors' => $user->getErrors(),
-                ]);
-            }
-
-            $this->setFailFlash(Craft::t('app', 'Couldn’t save user.'));
-
-            // Send the account back to the template
-            Craft::$app->getUrlManager()->setRouteParams([
-                $userVariable => $user,
-            ]);
-
-            return null;
+            return $this->asModelFailure(
+                $user,
+                Craft::t('app', 'Couldn’t save user.'),
+                $userVariable
+            );
         }
 
         // If this is a new user and email verification isn't required,
@@ -1406,13 +1390,13 @@ JS,
 
         if ($this->request->getAcceptsJson()) {
             $return = [
-                'success' => true,
                 'id' => $user->id,
             ];
             if ($returnCsrfToken && $generalConfig->enableCsrfProtection) {
                 $return['csrfTokenValue'] = $this->request->getCsrfToken();
             }
-            return $this->asJson($return);
+
+            return $this->asSuccess($return);
         }
 
         if ($isPublicRegistration) {
@@ -1480,7 +1464,7 @@ JS,
 
             Craft::error('There was an error uploading the photo: ' . $exception->getMessage(), __METHOD__);
 
-            return $this->asErrorJson(Craft::t('app', 'There was an error uploading your photo: {error}', [
+            return $this->asFailure(Craft::t('app', 'There was an error uploading your photo: {error}', [
                 'error' => $exception->getMessage(),
             ]));
         }
@@ -1546,17 +1530,9 @@ JS,
 
         $emailSent = Craft::$app->getUsers()->sendActivationEmail($user);
 
-        if ($this->request->getAcceptsJson()) {
-            return $this->asJson(['success' => $emailSent]);
-        }
-
-        if ($emailSent) {
-            $this->setSuccessFlash(Craft::t('app', 'Activation email sent.'));
-        } else {
-            $this->setFailFlash(Craft::t('app', 'Couldn’t send activation email. Check your email settings.'));
-        }
-
-        return $this->redirectToPostedUrl();
+        return $emailSent ?
+            $this->asSuccess(Craft::t('app', 'Activation email sent.')) :
+            $this->asFailure(Craft::t('app', 'Couldn’t send activation email. Check your email settings.'));
     }
 
     /**
@@ -1834,10 +1810,10 @@ JS,
         $this->requireAcceptsJson();
 
         if ($this->_verifyExistingPassword()) {
-            return $this->asJson(['success' => true]);
+            return $this->asSuccess();
         }
 
-        return $this->asErrorJson(Craft::t('app', 'Invalid password.'));
+        return $this->asFailure(Craft::t('app', 'Invalid password.'));
     }
 
     /**
@@ -1863,23 +1839,18 @@ JS,
         ]);
         $this->trigger(self::EVENT_LOGIN_FAILURE, $event);
 
-        if ($this->request->getAcceptsJson()) {
-            return $this->asJson([
+        return $this->asFailure(
+            $event->message,
+            data: [
                 'errorCode' => $authError,
-                'error' => $event->message,
-            ]);
-        }
-
-        $this->setFailFlash($event->message);
-
-        Craft::$app->getUrlManager()->setRouteParams([
-            'loginName' => $this->request->getBodyParam('loginName'),
-            'rememberMe' => (bool)$this->request->getBodyParam('rememberMe'),
-            'errorCode' => $authError,
-            'errorMessage' => $event->message,
-        ]);
-
-        return null;
+            ],
+            routeParams: [
+                'loginName' => $this->request->getBodyParam('loginName'),
+                'rememberMe' => (bool)$this->request->getBodyParam('rememberMe'),
+                'errorCode' => $authError,
+                'errorMessage' => $event->message,
+            ]
+        );
     }
 
     /**
@@ -1900,7 +1871,6 @@ JS,
         // If this was an Ajax request, just return success:true
         if ($this->request->getAcceptsJson()) {
             $return = [
-                'success' => true,
                 'returnUrl' => $returnUrl,
             ];
 
@@ -1908,7 +1878,7 @@ JS,
                 $return['csrfTokenValue'] = $this->request->getCsrfToken();
             }
 
-            return $this->asJson($return);
+            return $this->asSuccess(data: $return);
         }
 
         return $this->redirectToPostedUrl($userSession->getIdentity(), $returnUrl);
@@ -2214,7 +2184,7 @@ JS,
         ]));
 
         if ($this->request->getAcceptsJson()) {
-            return $this->asErrorJson('InvalidVerificationCode');
+            return $this->asFailure('InvalidVerificationCode');
         }
 
         // If they don't have a verification code at all, and they're already logged-in, just send them to the post-login URL
@@ -2319,20 +2289,15 @@ JS,
      */
     private function _handleSendPasswordResetError(array $errors, ?string $loginName = null): ?Response
     {
-        if ($this->request->getAcceptsJson()) {
-            /** @noinspection CallableParameterUseCaseInTypeContextInspection */
-            $errors = implode(', ', $errors);
+        $errorString = implode(', ', $errors);
 
-            return $this->asErrorJson($errors);
-        }
-
-        // Send the data back to the template
-        Craft::$app->getUrlManager()->setRouteParams([
-            'errors' => $errors,
-            'loginName' => $loginName,
-        ]);
-
-        return null;
+        return $this->asFailure(
+            $errorString,
+            errors: $errors,
+            routeParams: [
+                'loginName' => $loginName,
+            ]
+        );
     }
 
     /**
@@ -2364,6 +2329,6 @@ JS,
         $this->requirePostRequest();
         $ids = $this->request->getRequiredBodyParam('ids');
         Craft::$app->getAnnouncements()->markAsRead($ids);
-        return $this->asJson(['success' => true]);
+        return $this->asSuccess();
     }
 }